feat: 股票分析时点回溯(as_of)增强 + 量化图表重构 + i18n 去重补译
- astock-data as_of: 测试加 #[serial(asof)] 消除并发竞态，增强 vendor 降级日志
- stockAnalysisStore/types: as_of 日期接入分析流与决策卡片
- 量化回测: WalkForward 折叠/OOS 图、成交明细表、回测 Tab 重构
- rt-workflow: 节点级 ctx 变量、code_executor 隔离，per_node_exec_ctx_vars 测试
- dao/harness/entities: workflow_executions 仓储与 DTO 对齐
- i18n: 清理 10 国语言重复 key、补全中文泄漏与未翻译条目
- 时间旅行组件(AsOfDatePicker/ReplaySweep/TimeAnchorTour) 接入 as_of

+++ src-tauri/crates/entities/src/workflow_executions.rs
@@ -13,9 +13,10 @@ pub struct Model {
     pub input_params: Option<String>,
     pub output_result: Option<String>,
     pub node_executions: Option<String>,
-    // DDL 为 BIGINT（v100_consolidated.rs），必须用 i64；旧库该列实际也是
-    // BIGINT，用 i32 会导致 PG 上 INT8→INT4 解码失败。
-    pub total_time_ms: Option<i64>,
+    // 实际库列是 INT4（INTEGER）。迁移 DDL 虽声明 BIGINT，但当前运行库该列
+    // 实际为 INTEGER，用 i64 会在 PG 上触发 INT8→INT4 解码失败。按"实体适配
+    // 库列"纪律改用 i32（INT4）。若日后重建库使列变回 BIGINT，需同步改回 i64。
+    pub total_time_ms: Option<i32>,
     pub created_at: i64,
     pub updated_at: i64,
 }
